rename new image readers, add aditional documentation to deprecated image reader
