+++ org.knime.knip.io/src/org/knime/knip/io/nodes/imgreader2/readfromdialog/ImgReader2NodeFactory.xml
